tax-processing-pipeline-python/requirements.txt
typing
pyyaml

google-api-core
google-cloud-documentai
google-cloud-firestore

Flask
gunicorn
